Setup structure for triggering events and writing to disk
