+++ src/state_management.rs
@@ -1,5 +1,7 @@
 use std::time::Instant;
 
+use crate::units::Seconds;
+
 #[derive(Debug)]
 pub struct MonitorState(State);
 
@@ -11,8 +13,21 @@ enum State {
 }
 
 impl MonitorState {
-    // TODO Implement notifications
     pub fn new() -> Self {
         Self(State::Start)
     }
+
+    /// Updates the state and fires and returns an event if applicable
+    pub fn process_response(&self, response: &crate::ping::PingResponse) -> Option<Event> {
+        // TODO Handle state updates
+        None
+    }
+}
+
+#[derive(Debug)]
+pub enum Event {
+    ConnectionFailed(String),
+    ConnectionStillDown(Seconds),
+    ConnectionRestoredAfter(Seconds),
+    Error(String),
 }
